+++ app/(mainSite)/events/page.tsx
@@ -2,8 +2,14 @@ import React, { Suspense } from "react";
 
 import Footer from "@/components/Footer/Footer";
 import EventCards from "@/components/PageCard/EventCards";
-import { Box, Center, Grid, styled } from "@/styled-system/jsx";
+import { Box, Grid, styled } from "@/styled-system/jsx";
 import Loading from "../loading";
+import { getMetadata } from "@/lib/SEO";
+
+export const metadata = getMetadata({
+    title: "Events | NSC Computer Science Club",
+    url: "https://northcs.org/events",
+});
 
 /**
  * The events page containing all the events
@@ -13,7 +19,6 @@ import Loading from "../loading";
 const Events: React.FC = () => {
     return (
         <>
-            {/* <SEO url="https://northcs.org/events" /> */}
             <Box mt={0.5} pt={2} pb={2} bg="#95ca59">
                 <styled.h2
                     color="black"
